Implement the requested changes
